+++ src/bots/maintrait.rs
@@ -2,5 +2,5 @@ use gamecore::{GameBoard,State};
 
 pub trait Bot {
     fn new(side: State) -> Self;
-    fn get_move(&self, board: &GameBoard) -> (uint, uint);
+    fn get_move(&self, board: &GameBoard) -> (usize, usize);
 }
